# docs/openapi/schema.json
openapi: 3.0.3
info:
  title: SmartQueue API
  version: 1.0.0
  description: API for managing queues, tickets, and related services
paths:
...
  /api/notifications/preferences/:
    get:
      operationId: notifications_preferences_list
      tags:
      - notifications
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationPreference'
          description: ''
    post:
      operationId: notifications_preferences_create
      tags:
      - notifications
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreference'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NotificationPreference'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NotificationPreference'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreference'
          description: ''
...
components:
  schemas:
...
    ChannelTypeEnum:
      enum:
      - EM
      - SM
      - PU
      - WA
      - IA
      type: string
      description: |-
        * `EM` - Courriel
        * `SM` - SMS
        * `PU` - Push
        * `WA` - WhatsApp
        * `IA` - In-App
...
    NotificationPreference:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        channel_type:
          $ref: '#/components/schemas/ChannelTypeEnum'
        is_enabled:
          type: boolean
        quiet_hours_start:
          type: string
          format: time
          nullable: true
        quiet_hours_end:
          type: string
          format: time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        user:
          type: integer
          readOnly: true
      required:
      - channel_type
      - created_at
      - id
      - updated_at
      - user
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT